# backend/test/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/events": {
      "post": {
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Events"
        ]
      },
      "get": {
        "operationId": "EventsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by. Each endpoint documents its own allowed values.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "planning",
                "approval_pending",
                "approved",
                "live",
                "completed",
                "archived"
              ]
            }
          },
          {
            "name": "owner",
            "required": false,
            "in": "query",
            "description": "Owner user id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Free-text search across name, description and venue",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDateFrom",
            "required": false,
            "in": "query",
            "description": "Only events starting on or after this date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDateTo",
            "required": false,
            "in": "query",
            "description": "Only events starting on or before this date",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Events"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/status": {
      "patch": {
        "operationId": "EventsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Events"
        ]
      }
    },
    "/events/{id}/archive": {
      "patch": {
        "operationId": "EventsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Events"
        ]
      }
    },
...
    "/incidents": {
      "get": {
        "operationId": "IncidentsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Field to sort by. Each endpoint documents its own allowed values.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "event",
            "required": false,
            "in": "query",
            "description": "Filter by event id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignee",
            "required": false,
            "in": "query",
            "description": "Filter by assignee user id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "severity",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "critical",
                "high",
                "medium",
                "low"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "investigating",
                "mitigated",
                "resolved"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Incidents"
        ]
      }
    },
...
    "/analytics/overview": {
      "get": {
        "operationId": "AnalyticsController_getOverview",
        "parameters": [
          {
            "name": "event",
            "required": false,
            "in": "query",
            "description": "Scope the analytics to a single event. Event Managers are always scoped to their own events regardless of this value.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Analytics"
        ]
      }
    },
    "/analytics/workload": {
      "get": {
        "operationId": "AnalyticsController_getWorkload",
        "parameters": [
          {
            "name": "event",
            "required": false,
            "in": "query",
            "description": "Scope the analytics to a single event. Event Managers are always scoped to their own events regardless of this value.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Analytics"
        ]
      }
    },
    "/analytics/incidents": {
      "get": {
        "operationId": "AnalyticsController_getIncidentAnalytics",
        "parameters": [
          {
            "name": "event",
            "required": false,
            "in": "query",
            "description": "Scope the analytics to a single event. Event Managers are always scoped to their own events regardless of this value.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Analytics"
        ]
      }
    },
...
    "/analytics/approval-turnaround": {
      "get": {
        "operationId": "AnalyticsController_getApprovalTurnaround",
        "parameters": [
          {
            "name": "event",
            "required": false,
            "in": "query",
            "description": "Scope the analytics to a single event. Event Managers are always scoped to their own events regardless of this value.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Analytics"
        ]
      }
    },
    "/analytics/room-utilization": {
      "get": {
        "operationId": "AnalyticsController_getRoomUtilization",
        "parameters": [
          {
            "name": "event",
            "required": false,
            "in": "query",
            "description": "Scope the analytics to a single event. Event Managers are always scoped to their own events regardless of this value.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Analytics"
        ]
      }
    },
    "/analytics/scheduling-conflicts": {
      "get": {
        "operationId": "AnalyticsController_getSchedulingConflicts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Analytics"
        ]
      }
    }
  },
  "info": {
    "title": "EventOps Intelligence Platform API",
    "description": "Internal event-operations platform API documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "example": "jane@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongP@ssw0rd",
            "minLength": 8
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "jane@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongP@ssw0rd"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Annual Tech Summit"
          },
          "description": {
            "type": "string",
            "example": "A gathering of industry leaders"
          },
          "venue": {
            "type": "string",
            "example": "Grand Convention Center"
          },
          "startDate": {
            "type": "string",
            "example": "2026-10-01T09:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "example": "2026-10-03T18:00:00.000Z"
          }
        },
        "required": [
          "name",
          "venue",
          "startDate",
          "endDate"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Annual Tech Summit"
          },
          "description": {
            "type": "string",
            "example": "A gathering of industry leaders"
          },
          "venue": {
            "type": "string",
            "example": "Grand Convention Center"
          },
          "startDate": {
            "type": "string",
            "example": "2026-10-01T09:00:00.000Z"
          },
          "endDate": {
            "type": "string",
            "example": "2026-10-03T18:00:00.000Z"
          }
        }
      },
      "UpdateEventStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "planning",
              "approval_pending",
              "approved",
              "live",
              "completed",
              "archived"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}